Add opaque string generators and hegel_generate_string to the C ABI
hegel_string_generator_t handles are built by typed constructors (text
with codec/codepoint/category/include/exclude alphabets, regex with an
optional text-generator alphabet, email, url, domain), validate all
their parameters at construction time, are immutable and shareable
across test cases and threads, and are released with
hegel_string_generator_free. hegel_generate_string draws through a
handle and returns an engine-allocated, length-prefixed UTF-8 buffer
the caller frees with hegel_generate_string_result_free.

The regex, internet, and text-alphabet generation logic moves from the
schema interpreters to a typed native/draws module (along with
many_more/many_reject); the schema layer now delegates through thin
CBOR adapters. Compound engine-side string draws are wrapped in new
REGEX/EMAIL/URL/DOMAIN span labels.
