+++ hegel-c/src/native/draws/tlds-alpha-by-domain.txt
